patch 8.2.2845: MS-Windows: warning for signed/unsigned comparison
Problem:    MS-Windows: warning for signed/unsigned comparison.
Solution:   Add type cast.

+++ src/terminal.c
@@ -4341,13 +4341,13 @@ sync_shell_dir(VTermStringFragment *frag)
     char_u    *new_dir;
 
     // remove HOSTNAME to get PWD
-    while (*pos != '/' && offset < frag->len)
+    while (*pos != '/' && offset < (int)frag->len)
     {
         offset += 1;
         pos += 1;
     }
 
-    if (offset >= frag->len)
+    if (offset >= (int)frag->len)
     {
         semsg(_(e_failed_to_extract_pwd_from_str_check_your_shell_config),
 								    frag->str);
